+++ AbstractFactory.h
@@ -4,15 +4,3 @@
 struct AbstractFactory {
   static int bill_id;
 };
-
-struct DebitBillFactory : public AbstractFactory {
-  DebitBill CreateDebit();
-};
-
-struct CreditBillFactory : public AbstractFactory {
-  CreditBill CreateCredit();
-};
-
-struct DepositBillFactory : public AbstractFactory {
-  DepositBill CreateDeposit();
-};
