+++ CreditBill.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+#include "Bill.h"
+
+struct CreditBill : public Bill {
+  double Comission;
+};
