+++ CreditBillFactory.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "CreditBill.h"
+#include "AbstractFactory.h"
+
+struct CreditBillFactory : public AbstractFactory {
+  CreditBill CreateCredit();
+};
