+++ DebitBill.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+#include "Bill.h"
+
+struct DebitBill : public Bill{
+};
