+++ DebitBillFactory.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "DebitBill.h"
+#include "AbstractFactory.h"
+
+struct DebitBillFactory : public AbstractFactory {
+  DebitBill CreateDebit();
+};
