+++ DepositBillFactory.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "DepositBill.h"
+#include "AbstractFactory.h"
+
+struct DepositBillFactory : public AbstractFactory {
+  DepositBill CreateDeposit();
+};
