+++ bd_bills.csv
@@ -3,3 +3,5 @@
 4,1,0,200
 4,1488695700000001,1,0
 5,1488695700000002,0,0
+69,1488695700000001,1,0
+96,1488695700000002,-1,0
